lua: Autospec creation for version 5.3.4
Because src/Makefile already defines CFLAGS near the top of the file,
referencing CFLAGS within SYSCFLAGS, as specified in the patch, does not
substitute the intended value from the build environment.

Instead of patching src/Makefile, define MYCFLAGS (and also MYLIBS)
appropriately for the toplevel 'make linux' command. This properly
propogates CFLAGS from the build environment for the build.

Another side effect of this change is that the debuginfo package is
generated correctly now, as '-g' is one of the options from default
CFLAGS.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>